Move AIM to privacy from generic
